Html2Text::convert() can accept options, and OPT_FOOTER_URLS.
Html2Text::convert() can now accept an array as a second parameter, which allows passing in various options for the conversion process. This array is passed through to iterateOverNode().

Added an option (OPT_FOOTER_URLS) for relocating link URLs to the end of the document, replacing them with a numerical placeholder in the text. Makes the document much easier to read if there are a lot of long URLs in the HTML.