Mix Microsoft namespace and paragraph elements
Microsoft exchange emails often include HTML which is undocumented and non-standard. The <o:p> tags are handled as paragraphs by DOMDocument which results in huge amounts of line returns all mopped up by $output = preg_replace("/\n\n\n*/im", "\n\n", $output); that results in double line returns everywhere. To fix this the o:p elements should be removed and anything with a classname of MsoNormal (the standard classname in any Microsoft export or outlook for a paragraph that behaves like a line return) being changed to a line with a break afterwards. This makes emails display correctly while not impacting emails that do not have this encoding.